Updated figure captioning for latex and odf
Latex allows one to specify a shortened caption for inclusion in the
list of figures. Previously in MMD there was no way to specify this.
MMD does however allow a "title" attribute to be specified for HTML
figures. I added a mapping from this to the shortened caption for latex.

Also fixed odf caption output.